release-tool: fix set.sh Signed-off-by: DQ <dqstyle@gmail.com>

+++ build-files/kernel-zip/script/set.sh
@@ -37,9 +37,8 @@ PKG_IGNORE=/data/media/0/Synapse/pakage_list_ignore.txt
 PKG_SYSTEM=/data/media/0/Synapse/pakage_list_system.txt
 PKG_GOOGLE=/data/media/0/Synapse/pakage_list_google.txt
 PKG_AVAIL=/data/media/0/Synapse/pakage_list.txt
-#if [ ! -e $PKG_LIST ]; then
 cp -f /tmp/script/pakage_list_ignore.txt $PKG_IGNORE
 cp -f /tmp/script/pakage_list_system.txt $PKG_SYSTEM
 cp -f /tmp/script/pakage_list_google.txt $PKG_GOOGLE
-cp -f /tmp/script/pakage_list.txt $PKG_AVAIL
-#fi
+[ ! -f $PKG_AVAIL ] && cp -f /tmp/script/pakage_list.txt $PKG_AVAIL
+
